pyproject.toml
[project]
name = "calendar-reminders-mcp"
version = "0.1.0"
description = "MCP server for macOS Calendar and Reminders via EventKit"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyobjc-framework-eventkit>=12.1",
]

[project.scripts]
calendar-reminders-mcp = "src.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
